geo-16 исправлено межсервисное взаимодействие между user-service и server-service, добавлен контроллер Connection, Настроен сбор трафика юзера в БД

+++ backend/user-service/src/main/java/com/vpn/user/domain/entity/Connection.java
@@ -2,6 +2,8 @@ package com.vpn.user.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Entity
@@ -24,7 +26,7 @@ public class Connection {
     private Long deviceId;
 
     @Column(name = "server_id")
-    private Long serverId;
+    private Integer serverId;
 
     @Column(name = "bytes_sent")
     private Long bytesSent;
